Move protocol interfaces one level up
