+++ core/src/main/java/bisq/core/trade/TradeManager.java
@@ -52,11 +52,11 @@ import bisq.core.trade.model.bsq_swap.BsqSwapBuyerAsTakerTrade;
 import bisq.core.trade.model.bsq_swap.BsqSwapSellerAsMakerTrade;
 import bisq.core.trade.model.bsq_swap.BsqSwapSellerAsTakerTrade;
 import bisq.core.trade.model.bsq_swap.BsqSwapTrade;
+import bisq.core.trade.protocol.MakerProtocol;
 import bisq.core.trade.protocol.Provider;
+import bisq.core.trade.protocol.TakerProtocol;
 import bisq.core.trade.protocol.TradeProtocol;
 import bisq.core.trade.protocol.TradeProtocolFactory;
-import bisq.core.trade.protocol.bisq_v1.MakerProtocol;
-import bisq.core.trade.protocol.bisq_v1.TakerProtocol;
 import bisq.core.trade.protocol.bisq_v1.messages.InputsForDepositTxRequest;
 import bisq.core.trade.protocol.bisq_v1.model.ProcessModel;
 import bisq.core.trade.protocol.bsq_swap.BsqSwapMakerProtocol;
