+++ core/src/main/java/bisq/core/trade/protocol/SellerProtocol.java
@@ -15,17 +15,7 @@
  * along with Bisq. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package bisq.core.trade.protocol.bisq_v1;
+package bisq.core.trade.protocol;
 
-
-import bisq.core.trade.protocol.bisq_v1.messages.InputsForDepositTxRequest;
-
-import bisq.network.p2p.NodeAddress;
-
-import bisq.common.handlers.ErrorMessageHandler;
-
-public interface MakerProtocol {
-    void handleTakeOfferRequest(InputsForDepositTxRequest message,
-                                NodeAddress taker,
-                                ErrorMessageHandler errorMessageHandler);
+public interface SellerProtocol {
 }
