+++ core/src/main/java/bisq/core/trade/protocol/TakerProtocol.java
@@ -15,7 +15,12 @@
  * along with Bisq. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package bisq.core.trade.protocol.bisq_v1;
+package bisq.core.trade.protocol;
 
-public interface BuyerProtocol {
+public interface TakerProtocol {
+    void onTakeOffer();
+
+    enum TakerEvent implements FluentProtocol.Event {
+        TAKE_OFFER
+    }
 }
