+++ core/src/main/java/bisq/core/trade/protocol/bisq_v1/SellerAsTakerProtocol.java
@@ -21,6 +21,7 @@ package bisq.core.trade.protocol.bisq_v1;
 import bisq.core.offer.Offer;
 import bisq.core.trade.model.bisq_v1.SellerAsTakerTrade;
 import bisq.core.trade.model.bisq_v1.Trade;
+import bisq.core.trade.protocol.TakerProtocol;
 import bisq.core.trade.protocol.TradeMessage;
 import bisq.core.trade.protocol.bisq_v1.messages.CounterCurrencyTransferStartedMessage;
 import bisq.core.trade.protocol.bisq_v1.messages.DelayedPayoutTxSignatureResponse;
