+++ core/src/main/java/bisq/core/trade/protocol/bsq_swap/BsqSwapSellerAsTakerProtocol.java
@@ -39,7 +39,7 @@ import lombok.extern.slf4j.Slf4j;
 
 import static bisq.core.trade.model.bsq_swap.BsqSwapTrade.State.COMPLETED;
 import static bisq.core.trade.model.bsq_swap.BsqSwapTrade.State.PREPARATION;
-import static bisq.core.trade.protocol.bisq_v1.TakerProtocol.TakerEvent.TAKE_OFFER;
+import static bisq.core.trade.protocol.TakerProtocol.TakerEvent.TAKE_OFFER;
 import static com.google.common.base.Preconditions.checkNotNull;
 
 @Slf4j
